envia_app_scrapper/requirements.txt
# ========================================
# APP SCRAPER - Requirements
# Python 3.11.9+
# ========================================

# Web Scraping
playwright>=1.48,<2.0

# Google Sheets API
gspread==6.1.2
google-api-python-client==2.139.0
oauth2client==4.1.3

# Configuration
python-dotenv==1.0.1

# Async Support (incluido en Python 3.11+)
greenlet>=3.1.1
